requirements.txt
torch
torchvision
tqdm
anypath

python-dotenv
datasets==3.6.0

transformers

google-cloud-storage